Remove container from StatusLabel component and move their translations to the StatusTitle component
